requirements.txt
requests
flask>=2.0.0
flask-sqlalchemy
intersystems_iris @ https://github.com/intersystems-community/intersystems-irispython/releases/download/3.7.7/intersystems_iris-3.7.7-py3-none-any.whl
sqlalchemy_iris
iris-pex-embedded-python>=2.0.0
pytest
pandas
plotly